Allow enabling user namespaces with an annotation
This change enables user namespace support using the annotation
titus.netflix.com/host-users. Setting this annotation to false will
activate user namespace support. This is necessary to work around the
limitation that EKS does not enable user namespace support by default,
preventing us from using the PodSpec's HostUsers field to enable it. If
the annotation is not set, the behavior remains unchanged.

We will still respect the HostUsers field if it is set, using the
annotation as a fallback.

Signed-off-by: Jose Fernandez <josef@netflix.com>